# Improve JSON parsing performance