… TestTools/StreamChatTestTools/Fixtures/JSONs/BigChannelListPayload.json